requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
ecdsa==0.17.0
fastapi==0.65.2
fastapi-auth0==0.2.1
google-api-core==1.31.0
google-auth==1.33.0
google-cloud-core==1.7.1
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
h11==0.12.0
httptools==0.2.0
idna==2.10
imageio-ffmpeg==0.4.3
motor==2.4.0
ninja==1.10.0.post2
numpy==1.21.0
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
pyspng==0.1.0
python-dotenv==0.17.1
python-jose==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
scipy==1.7.0
six==1.16.0
starlette==0.14.2
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
